etl/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.1.0=gpu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
alabaster=0.7.12=pypi_0
alembic=1.7.7=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.2=pypi_0
astor=0.8.1=py38h06a4308_0
astroid=2.11.2=pypi_0
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
atomicwrites=1.4.0=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
binaryornot=0.4.4=pypi_0
black=22.3.0=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
blinker=1.4=py38h06a4308_0
blis=0.7.7=pypi_0
boruta=0.3=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.3.29=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=1.0.0=pypi_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=py38hd667e15_1
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cloudpickle=2.0.0=pypi_0
colorlover=0.3.0=pypi_0
cookiecutter=1.7.3=pypi_0
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cufflinks=0.17.3=pypi_0
cupti=10.1.168=0
cycler=0.11.0=pypi_0
cymem=2.0.6=pypi_0
databricks-cli=0.16.6=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diff-match-patch=20200713=pypi_0
dill=0.3.4=pypi_0
docker=5.0.3=pypi_0
docutils=0.17.1=pypi_0
entrypoints=0.4=pypi_0
fastjsonschema=2.15.3=pypi_0
flake8=4.0.1=pypi_0
flask=2.1.1=pypi_0
fonttools=4.32.0=pypi_0
frozenlist=1.2.0=py38h7f8727e_0
funcy=1.17=pypi_0
future=0.18.2=pypi_0
gast=0.4.0=pyhd3eb1b0_0
gensim=3.8.3=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
googleapis-common-protos=1.56.0=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.42.0=py38hce63b2e_0
gunicorn=20.1.0=pypi_0
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
htmlmin=0.1.12=pypi_0
idna=3.3=pyhd3eb1b0_0
imagehash=4.2.1=pypi_0
imagesize=1.3.0=pypi_0
imbalanced-learn=0.7.0=pypi_0
importlib-metadata=4.11.3=py38h06a4308_0
importlib-resources=5.6.0=pypi_0
inflection=0.5.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
intervaltree=3.1.0=pypi_0
ipykernel=6.13.0=pypi_0
ipython=7.32.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
isort=5.10.1=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=pypi_0
jeepney=0.8.0=pypi_0
jellyfish=0.9.0=pypi_0
jinja2=3.0.3=pypi_0
jinja2-time=0.2.0=pypi_0
joblib=1.0.1=pypi_0
jsonschema=4.4.0=pypi_0
jupyter-client=7.2.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.2.1=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
keyring=23.5.0=pypi_0
kiwisolver=1.4.2=pypi_0
kmodes=0.12.0=pypi_0
lazy-object-proxy=1.7.1=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libprotobuf=3.19.1=h4ff587b_0
libstdcxx-ng=9.3.0=hd4cf53a_17
lightgbm=3.3.2=pypi_0
llvmlite=0.37.0=pypi_0
lxml=4.8.0=pypi_0
mako=1.2.0=pypi_0
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mccabe=0.6.1=pypi_0
missingno=0.5.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mlflow=1.25.0=pypi_0
mlxtend=0.19.0=pypi_0
multidict=5.2.0=py38h7f8727e_2
multimethod=1.8=pypi_0
murmurhash=1.0.6=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.13=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
networkx=2.8=pypi_0
nltk=3.7=pypi_0
notebook=6.4.10=pypi_0
numba=0.54.1=pypi_0
numexpr=2.8.1=pypi_0
numpy=1.18.5=pypi_0
numpydoc=1.2.1=pypi_0
oauthlib=3.2.0=pyhd3eb1b0_0
openssl=1.1.1n=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pandas-profiling=3.1.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
patsy=0.5.2=pypi_0
pexpect=4.8.0=pypi_0
phik=0.12.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.1.0=pypi_0
pip=21.2.4=py38h06a4308_0
plac=1.1.3=pypi_0
platformdirs=2.5.1=pypi_0
plotly=5.7.0=pypi_0
pluggy=1.0.0=pypi_0
poyo=0.5.0=pypi_0
preshed=3.0.6=pypi_0
prettyconf=2.2.1=pypi_0
prometheus-client=0.14.1=pypi_0
prometheus-flask-exporter=0.20.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.19.1=py38h295c915_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
py4j=0.10.9.3=pypi_0
pyarrow=8.0.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycaret=2.3.10=pypi_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.9.0=pypi_0
pydocstyle=6.1.1=pypi_0
pyflakes=2.4.0=pypi_0
pygments=2.11.2=pypi_0
pyjwt=2.1.0=py38h06a4308_0
pyldavis=3.2.2=pypi_0
pylint=2.13.5=pypi_0
pyls-spyder=0.4.0=pypi_0
pynndescent=0.5.6=pypi_0
pyod=0.9.9=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.8=pypi_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyqtwebengine=5.15.5=pypi_0
pyqtwebengine-qt5=5.15.2=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
pyspark=3.2.1=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-lsp-black=1.2.0=pypi_0
python-lsp-jsonrpc=1.0.0=pypi_0
python-lsp-server=1.4.1=pypi_0
python-slugify=6.1.1=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.3.0=pypi_0
pyxdg=0.27=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
qdarkstyle=3.0.3=pypi_0
qstylizer=0.2.1=pypi_0
qtawesome=1.1.1=pypi_0
qtconsole=5.3.0=pypi_0
qtpy=2.0.1=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rope=1.0.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
rtree=1.0.0=pypi_0
scikit-learn=0.23.2=pypi_0
scikit-plot=0.3.7=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
secretstorage=3.3.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
snowballstemmer=2.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2=pypi_0
spacy=2.3.7=pypi_0
sphinx=4.5.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
spyder=5.3.0=pypi_0
spyder-kernels=2.3.0=pypi_0
sqlalchemy=1.4.35=pypi_0
sqlite=3.38.2=hc218d9a_0
sqlparse=0.4.2=pypi_0
srsly=1.0.5=pypi_0
statsmodels=0.13.2=pypi_0
tabulate=0.8.9=pypi_0
tangled-up-in-unicode=0.1.0=pypi_0
tenacity=8.0.1=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=gpu_py38h8a7d6ce_0
tensorflow-base=2.4.1=gpu_py38h29c2da4_0
tensorflow-datasets=4.5.2=pypi_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
tensorflow-gpu=2.4.1=h30adc30_0
tensorflow-metadata=1.7.0=pypi_0
termcolor=1.1.0=py38h06a4308_1
terminado=0.13.3=pypi_0
text-unidecode=1.3=pypi_0
textblob=0.17.1=pypi_0
textdistance=4.2.2=pypi_0
thinc=7.4.5=pypi_0
threadpoolctl=3.1.0=pypi_0
three-merge=0.1.1=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
tornado=6.1=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
ujson=5.2.0=pypi_0
umap-learn=0.5.2=pypi_0
unidecode=1.3.4=pypi_0
urllib3=1.26.8=pyhd3eb1b0_0
visions=0.7.4=pypi_0
wasabi=0.9.1=pypi_0
watchdog=2.1.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
wordcloud=1.8.1=pypi_0
wrapt=1.13.3=py38h7f8727e_2
wurlitzer=3.0.2=pypi_0
xz=5.2.5=h7b6447c_0
yapf=0.32.0=pypi_0
yarl=1.6.3=py38h27cfd23_0
yellowbrick=1.3.post1=pypi_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
